EPD-190: Fixed class cast exception

+++ epd-common/src/main/java/dk/dma/epd/common/prototype/gui/settings/CommonMapSettingsPanel.java
@@ -110,7 +110,7 @@ public class CommonMapSettingsPanel extends BaseSettingsPanel {
         lblDefaultMapScale.setBounds(103, 22, 113, 16);
         generalMapSettings.add(lblDefaultMapScale);
         
-        spinnerMaximumScale = new JSpinner(new SpinnerNumberModel(new Integer(0), null, null, new Integer(1)));
+        spinnerMaximumScale = new JSpinner(new SpinnerNumberModel(new Float(0), null, null, new Float(1)));
         spinnerMaximumScale.setBounds(16, 45, 75, 20);
         generalMapSettings.add(spinnerMaximumScale);
